requirements-simple.txt
fastapi==0.110.0
uvicorn==0.27.1
psycopg2-binary==2.9.9